pyproject.toml
[tool.poetry]
name = "flojoy-studio"
version = "0.1.18"
description = ""
authors = [
  "Jack Palmer <jp@flojoy.io>",
  "Joey Yu <joey@flojoy.io>",
  "Jeff Zhang <jeff@flojoy.io>",
  "Syed Mahbub <smahmed776@gmail.com>",
  "Dallas Strandell <dalstrande@outlook.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
flojoy = { path = "pkgs/flojoy", develop = true }
fastapi = "^0.103.2"
scikit-learn = "^1.3.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
debugpy = "^1.8.0"
opencv-python-headless = "^4.8.1.78"
pyserial = "^3.5"
zeroconf = "^0.115.2"
docstring-parser = "^0.15"
keyring = "^24.2.0"
qcodes = "^0.40.0"
pyvisa-py = "^0.7.0"
pyvisa = "^1.13.0"
pyusb = "^1.2.1"
tm-devices = {git = "https://github.com/tektronix/tm_devices.git", rev = "main"}
watchfiles = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
ruff = "^0.1.5"
typer = {extras = ["all"], version = "^0.9.0"}
docstring-parser = "^0.15"
python-frontmatter = "^1.0.1"


[tool.poetry.group.blocks.dependencies]
scipy = "^1.11.4"
scikit-image = "^0.22.0"
scikit-learn = "^1.3.2"
numpy = "^1.26.2"
mecademicpy = "^2.0.0"
boto3 = "^1.33.8"
pandas = {extras = ["parquet"], version = "^2.1.3"}
pyvisa-sim = "^0.6.0"
ticlib = "^0.2.2"
lxml = "^4.9.3"
labjackpython = "^2.1.0"
moto = "^4.2.11"
phidget22 = "^1.17.20231004"
pyserial = "^3.5"
pyyaml = "^6.0.1"
openpyxl = "^3.1.2"
opencv-python-headless = "^4.8.1.78"
sympy = "^1.12"
xlrd = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
